mySqrt

+++ mySqrt.cpp
@@ -10,7 +10,7 @@ public:
         int left = 0;
         int right = x;
         while (left <= right){
-           long long mid = (left + right) / 2;      // ให้มันคำนวณได้หลายครั้ง
+           long long mid = ((long long) left + right) / 2;      // ให้มันคำนวณได้หลายครั้ง    เปลี่ยนเป็น long long เพราะ overflow
             if (mid * mid < x){ 
                 left = mid + 1;           //move pointer to right
             } else if (mid * mid > x){
